+++ philander/serialbus_micropython.py
@@ -7,10 +7,11 @@ __author__ = "Oliver Maye"
 __version__ = "0.1"
 __all__ = ["_SerialBus_Micropython" ]
 
-from machine import I2C
+from machine import I2C, SPI
 
-from philander.serialbus import SerialBus
-from philander.sysfactory import SysProvider
+from philander.gpio import GPIO
+from philander.serialbus import SerialBus,SerialBusType, SPIMode
+from philander.sysfactory import SysFactory, SysProvider
 from philander.systypes import ErrorCode
 
     
@@ -21,13 +22,25 @@ class _SerialBus_Micropython( SerialBus ):
     def __init__(self):
         super().__init__()
         self.provider = SysProvider.MICROPYTHON
-        
+    
     def open( self, paramDict ):
         # Scan the parameters
         ret = super().open(paramDict)
         if (ret.isOk()):
             try:
-                self.bus = I2C( self.designator )
+                speed= paramDict["SerialBus.speed"]
+                if self.type == SerialBusType.I2C:
+                    self.bus = I2C( self.designator, freq=speed )
+                elif self.type == SerialBusType.SPI:
+                    mode = paramDict["SerialBus.SPI.mode"]
+                    pol = 0 if mode in [SPIMode.CPOL0_CPHA0, SPIMode.CPOL0_CPHA1] else 1
+                    ph  = 0 if mode in [SPIMode.CPOL0_CPHA0, SPIMode.CPOL1_CPHA0] else 1
+                    if paramDict["SerialBus.SPI.bitorder"].lower() == "MSB".lower():
+                        bit1 = SPI.MSB
+                    else:
+                        bit1 = SPI.LSB
+                    bpw  = paramDict["SerialBus.SPI.bpw"]
+                    self.bus = SPI( self.designator, baudrate=speed, polarity=pol, phase=ph, bits=bpw, firstbit=bit1 )
             except TypeError:
                 ret = ErrorCode.errInvalidParameter
         return ret
@@ -37,24 +50,58 @@ class _SerialBus_Micropython( SerialBus ):
         if (not self.bus is None) and hasattr(self.bus, "deinit"):
             self.bus.deinit()
         return ret
-    
+
+    def attach( self, device ):
+        result = ErrorCode.errOk
+        if (self.type == SerialBusType.SPI) and (device.pinCS is None):
+            result = ErrorCode.errFewData  # Need explicit CS pin designator!
+        if result.isOk():
+            result = super().attach( device )
+        return result
+
     def _readBytes( self, device, reg, num ):
         err = ErrorCode.errOk
-        try:
-            data = self.bus.readfrom_mem( device.address, reg, num )
-            data = int.from_bytes( data, "little" )
-        except OSError:
-            data = 0
-            err = ErrorCode.errLowLevelFail
+        data = 0
+        if self.type == SerialBusType.I2C:
+            try:
+                resp = self.bus.readfrom_mem( device.address, reg, num )
+                data = int.from_bytes( resp, "little" )
+            except OSError:
+                err = ErrorCode.errLowLevelFail
+        elif self.type == SerialBusType.SPI:
+            # Note that CS activation/de-activation is NOT done by hardware
+            device.pinCS.set( GPIO.LEVEL_LOW )
+            try:
+                self.bus.write( reg.to_bytes() )
+                resp = self.bus.read(num)
+                data = int.from_bytes( resp, "little" )
+            except OSError:
+                err = ErrorCode.errLowLevelFail
+            finally:
+                device.pinCS.set( GPIO.LEVEL_HIGH )
+        else:
+            err = ErrorCode.errNotSupported
         return data, err
 
     def _writeBytes( self, device, reg, data, num ):
         err = ErrorCode.errOk
-        try:
-            buf = data.to_bytes( num, "little" )
-            self.bus.writeto_mem( device.address, reg, buf )
-        except OSError:
-            err = ErrorCode.errLowLevelFail
+        if self.type == SerialBusType.I2C:
+            try:
+                buf = data.to_bytes( num, "little" )
+                self.bus.writeto_mem( device.address, reg, buf )
+            except OSError:
+                err = ErrorCode.errLowLevelFail
+        elif self.type == SerialBusType.SPI:
+            device.pinCS.set( GPIO.LEVEL_LOW )
+            try:
+                buf = reg.to_bytes( 1 ) + data.to_bytes( num, "little" )
+                self.bus.write( buf )
+            except OSError:
+                err = ErrorCode.errLowLevelFail
+            finally:
+                device.pinCS.set( GPIO.LEVEL_HIGH )
+        else:
+            err = ErrorCode.errNotSupported
         return err
 
     def readByteRegister( self, device, reg ):
@@ -77,37 +124,81 @@ class _SerialBus_Micropython( SerialBus ):
     
     def readBufferRegister( self, device, reg, length ):
         err = ErrorCode.errOk
-        try:
-            data = self.bus.readfrom_mem( device.address, reg, length )
-            data = list(data)
-        except OSError:
-            data = []
-            err = ErrorCode.errLowLevelFail
+        data = []
+        if self.type == SerialBusType.I2C:
+            try:
+                resp = self.bus.readfrom_mem( device.address, reg, length )
+                data = list(resp)
+            except OSError:
+                err = ErrorCode.errLowLevelFail
+        elif self.type == SerialBusType.SPI:
+            # Note that CS activation/de-activation is NOT done by hardware
+            device.pinCS.set( GPIO.LEVEL_LOW )
+            try:
+                self.bus.write([reg])
+                resp = self.bus.read(length)
+                data = list(resp)
+            except OSError:
+                err = ErrorCode.errLowLevelFail
+            finally:
+                device.pinCS.set( GPIO.LEVEL_HIGH )
+        else:
+            err = ErrorCode.errNotSupported
         return data, err
 
     def writeBufferRegister( self, device, reg, data ):
         err = ErrorCode.errOk
-        try:
-            self.bus.writeto_mem( device.address, reg, bytes(data) )
-        except OSError:
-            err = ErrorCode.errLowLevelFail
+        if self.type == SerialBusType.I2C:
+            try:
+                self.bus.writeto_mem( device.address, reg, bytes(data) )
+            except OSError:
+                err = ErrorCode.errLowLevelFail
+        elif self.type == SerialBusType.SPI:
+            device.pinCS.set( GPIO.LEVEL_LOW )
+            try:
+                buf = reg.to_bytes( 1 ) + bytes( data )
+                self.bus.write( buf )
+            except OSError:
+                err = ErrorCode.errLowLevelFail
+            finally:
+                device.pinCS.set( GPIO.LEVEL_HIGH )
+        else:
+            err = ErrorCode.errNotSupported
         return err
 
     def readBuffer( self, device, length ):
-        err = ErrorCode.errOk
-        try:
-            data = self.bus.readfrom( device.address, length )
-            data = list(data)
-        except OSError:
-            data = []
-            err = ErrorCode.errLowLevelFail
+        data, err = self.readWriteBuffer( device, length )
         return data, err
 
     def writeBuffer( self, device, buffer ):
-        err = ErrorCode.errOk
-        try:
-            self.bus.writeto( device.address, bytes(buffer) )
-        except OSError:
-            err = ErrorCode.errLowLevelFail
+        _, err = self.readWriteBuffer( device, 0, buffer )
         return err
     
+    def readWriteBuffer( self, device, inLength=0, outBuffer=None ):
+        err = ErrorCode.errOk
+        data = []
+        if self.type == SerialBusType.I2C:
+            try:
+                if (outBuffer is not None) and (len(outBuffer) > 0):
+                    self.bus.writeto( device.address, bytes(outBuffer) )
+                if inLength > 0:
+                    resp = self.bus.readfrom( device.address, inLength )
+                    data = list(resp)
+            except OSError:
+                err = ErrorCode.errLowLevelFail
+        elif self.type == SerialBusType.SPI:
+            device.pinCS.set( GPIO.LEVEL_LOW )
+            try:
+                if (outBuffer is not None) and (len(outBuffer) > 0):
+                    self.bus.write( bytes(outBuffer) )
+                if inLength > 0:
+                    resp = self.bus.read(inLength)
+                    data = list(resp)
+            except OSError:
+                err = ErrorCode.errLowLevelFail
+            finally:
+                device.pinCS.set( GPIO.LEVEL_HIGH )
+        else:
+            err = ErrorCode.errNotSupported
+        
+        return data, err